Commit_6

+++ database/migrations/2014_10_12_000000_create_users_table.php
@@ -21,7 +21,7 @@ class CreateUsersTable extends Migration{
          $table->string("phone_number")->unique();
          $table->string('e_mail')->unique();
          $table->string('password');
-         $table->bigInteger("role_id")->nullable();
+         $table->tinyInteger('role_id')->nullable();
          $table->timestamp('email_verified_at')->nullable();
          $table->rememberToken();
          $table->timestamps();
